requirements.txt
slackclient~=1.0
lxml~=3.6
python-decouple~=3.0